lambda/requirements.txt
# Lambda Dependencies (for runtime)
boto3>=1.34.0
pyTelegramBotAPI>=4.15.0
google-cloud-vision>=3.4.0
Pillow>=10.2.0
openai>=1.52.0
pydantic>=2.6.0
psycopg2-binary>=2.9.0
